Add my version of Doug Song's Python pcap library to the package for completeness.
Update all tests to look for the installed version of PCS, instead of
locally.  This can be undone by removing a single comment character.

Update the Makefile to correctly build PCS and the pcap.so no matter
what install or build target is selected.

Add Li-Wen Hsu and Tom McLaughlin as contributors.