Add pretty printing functions.
Make TCP correctly use its own offset field (number of 32 bit words
that the data is offset into the packet).

Add a rudimentary next() method to the TCP packet.

Change the use of next in IPv4 and IPv6 classes to correctly follow
the PCS convention.  next is no longer a field, only a method.