Bring in a change which makes TCP option parsing work.
Assume that a TLV contains itself, i.e. length field includes
the length of the type and length fields as well as the value.

Furthermore, assume that the length field encodes a length
measured in bytes, not bits. The rest of PCS assumes a field
width is measured in bits.