Rewrite the match facility totally.
Now it's really simple. By default, fields which are not initialized
by keyword arguments or by assignment are ignored. Otherwise, if a match
function for the field has not already been installed, we install a
default compare function which lives in each Field class as a class method.

The expect.py example is updated to show how easy using this is.