Add calc_length() support to most packets.
Packets and Chains now know how to fill out their own length fields
with a little help from Chain. This has to be done explicitly because
we want to retain the capability of representing badly formed packets as
Python objects.  SCTP is excluded for the time being.
This makes picking up and using PCS 'on-the-go' vastly less tedious.

While we're at it fix a bug in IGMPv3 query decoding.